spotify-custom-wakeword/requirements.txt
absl-py==1.4.0
altair==4.2.2
appdirs==1.4.4
astunparse==1.6.3
attrs==22.2.0
audioread==3.0.0
bitsets==0.8.4
blinker==1.5
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
concepts==0.9.2
conda==23.1.0
conda-content-trust==0.1.3
conda-package-handling==2.0.2
conda_package_streaming==0.7.0
contourpy==1.0.7
cryptography==38.0.4
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
features==0.5.12
fileconfig==0.6.1
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.53.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.1.0
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.39.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.0.4
model-profiler==1.1.8
msgpack==1.0.5
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pip==22.3.1
pluggy==1.0.0
pooch==1.6.0
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.4
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-speech-features==0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.2.2
scipy==1.10.1
semver==2.13.0
setuptools==65.6.3
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.4
streamlit==1.20.0
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.14
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.2.3
wheel==0.37.1
wrapt==1.14.1
zipp==3.15.0
zstandard==0.18.0
